fix: Update onboarding experience
