+++ Composer/packages/client/__tests__/components/appSettings.test.tsx
@@ -18,13 +18,14 @@ describe('<AppSettings /> & <ElectronSettings />', () => {
   });
 
   it('should render the user settings page', () => {
-    const { getByText } = renderWithRecoil(<AppSettings />, ({ set }) => {
+    const { getByText, getAllByText } = renderWithRecoil(<AppSettings />, ({ set }) => {
       set(onboardingState, {
         coachMarkRefs: {},
         complete: false,
       });
     });
     // there are 2 onboarding texts
+    getAllByText('Onboarding');
     getByText('Property editor preferences');
     getByText('Application Updates');
   });
