+++ Composer/packages/client/src/Onboarding/WelcomeModal/Expanded/styles.ts
@@ -2,12 +2,12 @@
 // Licensed under the MIT License.
 
 import { css } from '@emotion/core';
-import { NeutralColors, SharedColors } from '@uifabric/fluent-theme/lib/fluent/FluentColors';
+import { NeutralColors, SharedColors } from '@uifabric/fluent-theme';
 
 export const buttonStyle = css`
   position: absolute;
-  right: 0px;
-  top: 0px;
+  right: -24px;
+  top: -24px;
 `;
 
 export const contentStyle = css`
