+++ Composer/packages/client/src/Onboarding/WelcomeModal/WelcomeModal.tsx
@@ -3,18 +3,40 @@
 
 import React, { useContext } from 'react';
 import { Modal } from 'office-ui-fabric-react/lib/Modal';
+import { FluentTheme } from '@uifabric/fluent-theme';
+import { IModalStyles } from 'office-ui-fabric-react/lib/Modal';
 
-import OnboardingContext from '../OnboardingContext';
+import { useOnboardingContext } from '../OnboardingContext';
 
 import CollapsedWelcomeModal from './Collapsed/CollapsedWelcomeModal';
 import ExpandedWelcomeModal from './Expanded/ExpandedWelcomeModal';
-import { collapsedStyles, expandedStyles } from './style';
+
+const collapsedStyles: Partial<IModalStyles> = {
+  main: {
+    backgroundColor: FluentTheme.palette.themePrimary,
+    bottom: '30px',
+    minHeight: '55px',
+    color: 'white',
+    paddingLeft: '15px',
+    position: 'absolute',
+    right: '20px',
+  },
+};
+
+const expandedStyles: Partial<IModalStyles> = {
+  main: {
+    bottom: '30px',
+    padding: '15px',
+    position: 'absolute',
+    right: '30px',
+  },
+};
 
 const WelcomeModal = () => {
   const {
     actions: { exit },
     state: { complete, hideModal, minimized },
-  } = useContext(OnboardingContext);
+  } = useOnboardingContext();
 
   return !(complete || hideModal) ? (
     <Modal
