--- Composer/packages/client/src/Onboarding/palette.ts
@@ -1,29 +1,0 @@
-// Copyright (c) Microsoft Corporation.
-// Licensed under the MIT License.
-
-import { IPalette } from 'office-ui-fabric-react/lib/Styling';
-
-export const palette: Partial<IPalette> = {
-  themePrimary: '#5c2e91',
-  themeLighterAlt: '#040206',
-  themeLighter: '#0f0717',
-  themeLight: '#1c0e2c',
-  themeTertiary: '#381c57',
-  themeSecondary: '#522980',
-  themeDarkAlt: '#693d9c',
-  themeDark: '#7c53ac',
-  themeDarker: '#9a78c2',
-  neutralLighterAlt: '#f8f8f8',
-  neutralLighter: '#f4f4f4',
-  neutralLight: '#eaeaea',
-  neutralQuaternaryAlt: '#dadada',
-  neutralQuaternary: '#d0d0d0',
-  neutralTertiaryAlt: '#c8c8c8',
-  neutralTertiary: '#0f0f0f',
-  neutralSecondary: '#1f1f1f',
-  neutralPrimaryAlt: '#2d2d2d',
-  neutralPrimary: '#333333',
-  neutralDark: '#646464',
-  black: '#8d8d8d',
-  white: '#ffffff',
-};
