+++ Composer/packages/client/src/constants.ts
@@ -410,4 +410,4 @@ export const runtimeOptions: IDropdownOption[] = [
   { key: 'azureFunctions', text: 'Azure Functions' },
 ];
 
-export const onboardingDisabled = true;
+export const onboardingDisabled = false;
